fix: multiple popups in login screen on device rotation
Earlier, rotating the device would cause multiple popups to be shown if
the user didn't dismiss the earlier ones. This was because the popup was
shown whenever onStart() of the LoginFragment was called. This is called
on device rotation, leading to this bug.

Now, a boolean tracks whether the popup has been shown before.

Fixes: #1066