fix: make app exit on back press on first-launch in events screen (#1098)
Earlier, if the user selected the "Use my current location" option in
the welcome screen and pressed the back button in the newly opened
events screen, he would see the welcome screen slide in partially for an
instant. Then, he would be shown the events screen again. Thus, the back
button wasn't working.

This was because the welcome screen remained on the backstack.

This commit fixes the issue and implements conditional navigation to the
welcome screen like Google's docs recommend.

Fixes: #992