chore: make two flavor specific Koin modules and split WelcomeFragment
WelcomeFragment has been split into a fdroid and a playstore variant.
This has been done because the fdroid version doesn't need a ViewModel
and any location services but the playstore version does.

The fdroid flavor had a redundant ViewModel and it's code was confusing
because the ViewModel was used for entirely different purposes in the
two flavors.

Making two flavor specific modules(one for fdroid and one for playstore)
allowed us the split the WelcomeFragment itself.

This has the benefit that the code in WelcomeFragment now reflects the
features of the flavor and the redundant ViewModel is no longer needed.

Fixes: #1082