fix: Rotating screen causes old snackbar to be shown again
Earlier, configuration changes would trigger the last snackbar to show
up again. This was unfortunate as snackbars are used to represent
transient things.

This was happening because MutableLiveData was being used instead of
SingleLiveEvent.

Not all ViewModels had this issue as some were already using
SingleLiveEvent.

This commit fixes the issue.

Fixes: #1127