Summary Change: fix bug #1150
Details: I checked if the view is already initialized. If it is, I reuse the view of search result, not reload the whole thing again. I also realized that the events fragment have the same problem so I make the same change. Besides, I prevent the changes in action bar

fix: #1150