Summary: Fix#1150 - After coming back from result detail, search result come back to the top, not current position
Details: I checked if the view is already initialized. If it is, I reuse the view of search result, not reload the whole thing again. I also realized that the events fragment have the same problem so I make the same change. I also set the action bar first before return the rootView

fix: #1150