fix: After coming back from result detail, search result come back to the top, not current position #1150
Details:
I checked if the view is already initialized and if a new search is not called. If it is, I reuse the view of the search result, not reload the whole thing again. I also realized that the events fragment have the same problem so I make the same change.