Summary Change: fix data loss on rotating EditProfileFragment
Details:
In here, I deleted the use of EditUserViewModel because it is only used to fetch the data in the beginning and set the edit text to original firstname/lastname . So whenever an user rotates the screen, a ViewModel is created again and fetch the data again -> editText follows those changes and keep reseting the text to the original firstname, lastname.

Instead, I initially pass the data from ProfileFragment so that we don't have to fetch the data from the server again. Text in the EditText is automatically retained now.

This commit fix the issue 1188