Summary Changes: Add remove favorite event animation
Details:
There is a problem in the FavoriteFragment: The way it updates the favorite event list is wrong. Basically, when an event is clicked on the favorite icon, it is not removed from the adapter list but only from the database by the viewModel, and then the viewModel notices the changes from the event list in the viewModel and fetch the whole list all over again.

I fixed this by making sure the viewModel only fetch the list once, remove the favorite event properly from the adapter and use notifyItemRemoved(), not notifyItemChanged() to make the animations.

This resolves issue 1285