chore: Move flask filter logic from ViewModels to Services
Earlier, ViewModels were building flask filter queries themselves and
passing them to Services.

This was not good because this leaked the data retrieval logic into the
ViewModels. Only the Services should be concerned with data retrieval
and they should abstract out that logic.

This change moves the logic into Services where it belongs.

Fixes: #1344