Summary Change: Add Recent Search Suggestion to SearchFragment
Details:
The activity is configured with Action SEARCH with a xml file "searchable" to get the assistance from Android System on search. Recent query is stored and received by SearchRecentSuggestionsProvider() object.

Besides, I also fix the bug to make close icon expanding to full view in the action bar

Fix: #575