Summary Changes: Retain fragments on bottom navigation on nav controller
Details:
An Advanced Navigation workaround from Google is used to retain all of the fragments. Each of the item in bottom navigation is retained as navigation graph (5 navigation graphs - Events, Search, Likes, Tickets, Profile).

Some of the changes are made for proper navigation between fragments from different navigation graph (Example: booking order finished in EventsFragment - Events navigation graph to seeing ticket in OrderUnderUserFragment - Tickets navigation graph)

In LoginFragment, if (loginViewModel.isLoggedIn()) popBackStack() is moved to onActivityCreated as popBackStack() function require rootView to be created.

In SearchFragment, queryListener is moved to onCreateView is because onPrepareOptionMenu is not called everytime we navigate to SearchFragment by bottom navigation, only the first time.

In SearchLocationFragment, fragmentManager is replaced by childFragmentManager because PlaceAutocompleteFragment is a fragment used inside a fragment, so using fragmentManager would produce "No view found for id 0x7f090065 for fragment PlaceAutocompleteFragment

Fixes: #1399