Summary Change: Add sponsors section
Details:
Add sponsors section
LATER: I will make another PR to make a sponsor fragment to display the detail information of the sponsors. It is weird that when I fetched the sponsors, the number of sponsors sent to the app can go up to 600, I don't know if someone really made those much data or are there any problems in the server side.

Fixes: #1651